requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
python-dotenv==1.0.1
requests==2.32.3
shiboken6==6.8.0.2
urllib3==2.2.3
